# Go Framework/web-server/docs/swagger.yaml
basePath: /
definitions:
  models.File:
    properties:
      id:
        type: integer
      total_letters:
        type: integer
      total_lines:
        type: integer
      total_spaces:
        type: integer
      total_special_char:
        type: integer
      total_words:
        type: integer
      user_id:
        type: integer
    type: object
  models.Login:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.Users:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: REST API built with Gin
  title: Web Server API
  version: "1.0"
paths:
  /count:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file and get word, letter, space, line, and special character
        counts
      parameters:
      - description: Document ID
        in: formData
        name: id
        required: true
        type: integer
      - description: File to analyze
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.File'
        "400":
          description: Invalid input or file
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Analyze a file
      tags:
      - analyzer
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.Login'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid login request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid email or password
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /result/{Id}:
    get:
      description: Retrieve a single analyzed file record by its ID for the authenticated
        user
      parameters:
      - description: Record ID
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.File'
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Document not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get result by ID
      tags:
      - analyzer
    patch:
      consumes:
      - application/json
      description: Update an analyzed file record by its ID for the authenticated
        user
      parameters:
      - description: Record ID
        in: path
        name: Id
        required: true
        type: integer
      - description: Updated file data
        in: body
        name: file
        required: true
        schema:
          $ref: '#/definitions/models.File'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.File'
        "400":
          description: Invalid ID or body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Record not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update result by ID
      tags:
      - analyzer
  /results:
    get:
      description: Retrieve all analyzed file records for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.File'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all results
      tags:
      - analyzer
  /signup:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Users'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid signup request or email
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User registration
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"